feat(refactor B): port logic & imports to ai_org_backend
