+++ backend/ai_org_backend/alembic/env.py
@@ -1,7 +1,6 @@
 # backend/ai_org_backend/alembic/env.py
 
 from logging.config import fileConfig
-from sqlalchemy import engine_from_config, pool
 from alembic import context
 
 # Importiere die SQLModel-MetaData!
