--- backend/ai_org_backend/alembic/versions/.gitkeep
@@ -1,1 +1,0 @@
-
