+++ backend/tests/test_smoke.py
@@ -1,0 +1,2 @@
+def test_placeholder():
+    assert True
